+++ lib/mongoid/errors/validations.rb
@@ -8,7 +8,12 @@ module Mongoid
     # @example Create the error.
     #   Validations.new(person.errors)
     class Validations < MongoidError
+      attr_reader :document
+      alias :record :document
+
       def initialize(document)
+        @document = document
+
         super(
           compose_message(
             "validations",